Removed unnused function and changed some logic around quantities
Removed the 'reduceItemQuantity', since another function does the same job with more and better functionality. Also changed the process with item's quantities, differentiating between a item's quantity, i.e how many examplars there are of it, and it's quantity in the database. Also added extra routes for the missing functions already created.